% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPfates.R
\name{GPfates}
\alias{GPfates}
\title{Run GPfates}
\usage{
GPfates(counts, nfates = 1, ndims = 2, log_expression_cutoff = 2,
  min_cells_expression_cutoff = 2)
}
\arguments{
\item{counts}{The non-normalised expression counts}

\item{nfates}{The number of fates}

\item{ndims}{The number of dimensions}

\item{log_expression_cutoff}{The log expression cutoff}

\item{min_cells_expression_cutoff}{The min expression cutoff}
}
\description{
Run GPfates
}
